+++ Notepad2e.vcxproj
@@ -326,6 +326,7 @@
     <ClCompile Include="src\Extension\EditHelperEx.cpp" />
     <ClCompile Include="src\Extension\HLSelection.c" />
     <ClCompile Include="src\Extension\InlineProgressBarCtrl.c" />
+    <ClCompile Include="src\Extension\MainWndHelper.c" />
     <ClCompile Include="src\Extension\StrToHex.c" />
     <ClCompile Include="src\Extension\tinyexpr\tinyexpr.c" />
     <ClCompile Include="src\Extension\Utils.c" />
@@ -403,6 +404,7 @@
     <ClInclude Include="src\Extension\EditHelperEx.h" />
     <ClInclude Include="src\Extension\HLSelection.h" />
     <ClInclude Include="src\Extension\InlineProgressBarCtrl.h" />
+    <ClInclude Include="src\Extension\MainWndHelper.h" />
     <ClInclude Include="src\Extension\SciCall.h" />
     <ClInclude Include="src\Extension\StrToHex.h" />
     <ClInclude Include="src\Extension\tinyexpr\tinyexpr.h" />
